# 테스트용 사냥터 기능 추가, 스킬 발동 수정
## UnitSpawnTester
1. 몬스터 자동 스폰
    - 몬스터는 유닛(용병) 클래스를 임시 사용
    - n초마다 n마리의 몬스터 소환 (아래 진영 포탈)
    - 임의로 소환 가능 (아래 진영 포탈)

2. 유닛 소환
    - 소환할 때 테스트용 데이터에서 랜덤값을 뽑아서 스탯을 결정
    - 위쪽 진영 포탈에서 소환

## UnitSelectTest
- 2d 콜라이더와 IPointerDownHandler로 클릭 감지

## 유닛 사냥터AI 세부 구현
- 스킬 사용 상태 추가
- 각 상태의 우선순위 정리

## Skill
- 스킬 사용 조건 검사와 동시에 사용하던 것 변경
    - 사용은 다른 부분에서 처리 (유닛 AI 에서)

+++ Assets/Scripts/Utilities/SyncedTimeSetting.cs
@@ -61,5 +61,5 @@ public class SyncedTimeSetting : ScriptableObject
         HOUR
     }
     public SYNC_INTERVAL_TYPE syncIntervalType = SYNC_INTERVAL_TYPE.MINUTE;
-    public float syncInterval = 30f;
+    public int syncInterval = 30;
 }
